% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MplusTrees.R
\name{MplusTrees}
\alias{MplusTrees}
\title{Mplus model recursive partitioning trees}
\usage{
MplusTrees(script, data, rPartFormula, group = ~id, parallel = FALSE,
  control = rpart.control())
}
\arguments{
\item{script}{An MplusAutomation script file}

\item{data}{Dataset that is specified in the script}

\item{rPartFormula}{Formula of the form ~ variable names}

\item{group}{What variable is an id variable. If not specified
an id variable is created for each row.}

\item{parallel}{Whether to parallelize each node (max two).}

\item{control}{Control object for rpart.}
}
\description{
Mplus model recursive partitioning trees
}
